requirements.txt
Django==3.2.13
djangorestframework==3.11.0
psycopg2==2.9.3
djangorestframework-simplejwt